requirements.txt
streamlit>=1.30.0
pymupdf==1.23.5  # For PDF processing (fitz)
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0  # For loading environment variables